Begin moving user analytics functionality to separate composable
