+++ apps/web/src/pages/overview/components/BreakdownChart.vue
@@ -1,11 +1,12 @@
 <script lang="ts" setup>
 import LineChartJS from '@/components/charts/LineChartJS.vue'
-import { onMounted, ref, watch} from 'vue'
+import { onMounted, ref, toValue, watch} from 'vue'
+import useAnalytics from '@/composables/analytics'
 import useUser from '@/composables/user'
 import useScreenDimensions from '@/composables/screenDimensions'
-import { AnalyticsData, ProviderString } from '@casimir/types'
+import { AnalyticsData, ProviderString, UserAnalyticsData } from '@casimir/types'
 
-const { currentStaked, finishedComputingUerAnalytics, stakingRewards, totalWalletBalance, user, userAnalytics } = useUser()
+const { currentStaked, stakingRewards, totalWalletBalance, user } = useUser()
 const { screenWidth } = useScreenDimensions()
 
 const chardId = ref('cross_provider_chart')
@@ -51,25 +52,26 @@ const formatLegendLabel = (address: string) => {
   return (account? account.walletProvider : 'Unknown') + ' (' + start + middle + end + ')'
 }
 
-const setChartData = () => {
+const setChartData = (userAnalytics: UserAnalyticsData) => {
+  console.log('user analytics in BreakdownChart', userAnalytics)
   let labels
   let data: Array<AnalyticsData> = []
   switch (selectedTimeframe.value) {
     case '1 month':
-      labels = userAnalytics.value.oneMonth.labels
-      data = userAnalytics.value.oneMonth.data as any
+      labels = userAnalytics.oneMonth.labels
+      data = userAnalytics.oneMonth.data as any
       break
     case '6 months':
-      labels = userAnalytics.value.sixMonth.labels
-      data = userAnalytics.value.sixMonth.data as any
+      labels = userAnalytics.sixMonth.labels
+      data = userAnalytics.sixMonth.data as any
       break
     case '12 months':
-      labels = userAnalytics.value.oneYear.labels
-      data = userAnalytics.value.oneYear.data as any
+      labels = userAnalytics.oneYear.labels
+      data = userAnalytics.oneYear.data as any
       break
     case 'historical':
-      labels = userAnalytics.value.historical.labels
-      data = userAnalytics.value.historical.data as any
+      labels = userAnalytics.historical.labels
+      data = userAnalytics.historical.data as any
       break
     
     default:
@@ -93,21 +95,26 @@ const setChartData = () => {
   }
 }
 
-onMounted(async () => {
-    setChartData()
+
+const {userAnalytics, updateAnalytics } = useAnalytics()
+
+onMounted(() => {
+  setChartData(userAnalytics.value as UserAnalyticsData)
 })
 
 watch(selectedTimeframe, () => {
-  setChartData()
+  setChartData(userAnalytics.value as UserAnalyticsData)
+})
+
+watch(userAnalytics, () => {
+  setChartData(userAnalytics.value as UserAnalyticsData)
 })
 
 watch(user, async () => {
-    setChartData()
+  await updateAnalytics()
+  setChartData(userAnalytics.value as UserAnalyticsData)
 })
 
-watch(finishedComputingUerAnalytics, async () => {
-    setChartData()
-})
 </script>
 
 <template>
